refactored helpers into utils/
