+++ src/utils/models/tickets/index.js
@@ -1,0 +1,2 @@
+export * from './sortTickets'
+export * from './getAllBookedTickets'
