+++ src/utils/models/tickets/sortTickets.js
@@ -1,0 +1,8 @@
+/**
+ * Sorts tickets by their display order
+ * @param {*} tickets
+ */
+export const sortTickets = tickets =>
+  tickets.sort(
+    (ticketA, ticketB) => ticketA.displayOrder - ticketB.displayOrder
+  )
